Read templates from a templates file.
This allows tokens to be replaced in templates in addition to the file
names.

Support for using snippets has been removed as part of this. As there is
no API to programmatically read snippets, it is not possible to do token
replacement in them. The Emmet plugin does something similar.